Move `browser/browser.py` to `browser/server.py`
Suggested by @rdswift & @phw
